requirements.txt
Django==2.0.4
django-filters==0.2.1
djangorestframework==3.8.2
Markdown==2.6.11
pytz==2018.4
